Move navigation blocker to Editor component
