+++ ui/src/pages/namespace/Explorer/Endpoint/EndpointEditor/index.tsx
@@ -88,6 +88,11 @@ const EndpointEditor: FC<EndpointEditorProps> = ({ data }) => {
                     options={{
                       readOnly: true,
                     }}
+                    navigationWarning={
+                      isDirty
+                        ? t("components.blocker.unsavedChangesWarning")
+                        : null
+                    }
                   />
                 </Card>
               </div>
